Null safe migration of model library

+++ apps/payouts/lib/src/model/foundation.dart
@@ -1,11 +1,9 @@
-// @dart=2.9
-
 import 'dart:math' as math;
 
 /// Rounds the specified [value] to the specified number of significant digits.
 ///
 /// For example, `roundToSignificantDigits(1.234567, 2)` will return `1.23`.
 double roundToSignificantDigits(double value, int significantDigits) {
-  int rounder = math.pow(10, significantDigits);
+  int rounder = math.pow(10, significantDigits) as int;
   return (value * rounder).roundToDouble() / rounder;
 }
